requirements.txt
streamlit
openai
langchain
vaderSentiment
python-dotenv
matplotlib
pandas
